Add EPMConnector
