+++ EPMConnector/Client.cs
@@ -1,14 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Net.Sockets;
-using System.Threading.Tasks;
 using System.Threading;
 using Eleon.Modding;
 
 
-namespace ENRC.client
+namespace EPMConnector
 {
     public class Client
     {
@@ -18,26 +14,33 @@ namespace ENRC.client
         public event Action<ModProtocol.Package> GameEventReceived;
 
         volatile ModProtocol client;
-        const string cIP = "127.0.0.1";
-        const int cPort = 12345;
+
+        string gameServerIp;
+        int gameServerPort;
 
         public Client()
         {
+        }
+
+        public void Connect(string ipAddress, int port)
+        {
+            this.gameServerIp = ipAddress;
+            this.gameServerPort = port;
             connectToServerThread = ModThreadHelper.StartThread(ThreadConnectToServer, System.Threading.ThreadPriority.Lowest);
         }
 
         private void ThreadConnectToServer(ModThreadHelper.Info ti)
         {
-            ClientMessages(string.Format("ModInterface: Started connection thread. Connecting to {0}:{1}", cIP, cPort));
+            ClientMessages(string.Format("ModInterface: Started connection thread. Connecting to {0}:{1}", this.gameServerIp, this.gameServerPort));
             while (!ti.eventRunning.WaitOne(0))
             {
                 if (client == null)
                 {
                     try
                     {
-                       TcpClient tcpClient = new TcpClient(cIP, cPort);
-                       client = new ModProtocol(tcpClient, PackageReceivedDelegate, DisconnectedDelegate);
-                       ClientMessages("ModInterface: Connected with " + client);
+                        TcpClient tcpClient = new TcpClient(this.gameServerIp, this.gameServerPort);
+                        client = new ModProtocol(tcpClient, PackageReceivedDelegate, DisconnectedDelegate);
+                        ClientMessages("ModInterface: Connected with " + client);
                     }
                     catch (SocketException)
                     {
