+++ Empyrion Mod/EPM.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Eleon.Modding;
+using EPMConnector;
 
 public class EPM : ModInterface {
