+++ Empyrion Mod/EPM.csproj
@@ -38,14 +38,18 @@
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup>
     <None Include="common\ModInterface.cs" />
-    <Compile Include="common\ModProtocol.cs" />
     <Compile Include="common\ModTCPServer.cs" />
-    <Compile Include="common\ModThreadHelper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="EPM.cs" />
     <Compile Include="helper classes\ModLog.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="UnityEngine.dll" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\EPMConnector\EPMConnector.csproj">
+      <Project>{ae40b2d3-69ef-4d5e-b1fb-38eece976811}</Project>
+      <Name>EPMConnector</Name>
+    </ProjectReference>
   </ItemGroup>
 </Project>
