+++ Empyrion Mod/common/ModTCPServer.cs
@@ -4,6 +4,7 @@ using System.Net;
 using System.Net.Sockets;
 using System.Threading;
 using Eleon.Modding;
+using EPMConnector;
 
 public class ModTCPServer {
