+++ EmpyrionModSample.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Empyrion Network Relay Clie
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeathMessagesModule", "DeathMessenger\DeathMessagesModule.csproj", "{38D13EF0-A11E-4F94-8139-D0680C590C34}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EPMConnector", "EPMConnector\EPMConnector.csproj", "{AE40B2D3-69EF-4D5E-B1FB-38EECE976811}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{38D13EF0-A11E-4F94-8139-D0680C590C34}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{38D13EF0-A11E-4F94-8139-D0680C590C34}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{38D13EF0-A11E-4F94-8139-D0680C590C34}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AE40B2D3-69EF-4D5E-B1FB-38EECE976811}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AE40B2D3-69EF-4D5E-B1FB-38EECE976811}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AE40B2D3-69EF-4D5E-B1FB-38EECE976811}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AE40B2D3-69EF-4D5E-B1FB-38EECE976811}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
